+++ packages/frameworks/angular/src/dialog/dialog-close-button.component.ts
@@ -6,7 +6,7 @@ import {X} from "lucide-angular"
 
 import {CoreDialogCloseTriggerDirective} from "@qualcomm-ui/angular-core/dialog"
 import {provideIcons} from "@qualcomm-ui/angular-core/lucide"
-import {useInlineIconButtonApi} from "@qualcomm-ui/angular/inline-icon-button"
+import {useIconButtonApi} from "@qualcomm-ui/angular/button"
 import {mergeProps} from "@qualcomm-ui/utils/merge-props"
 
 import {useQdsDialogContext} from "./qds-dialog-context.service"
@@ -24,10 +24,10 @@ import {useQdsDialogContext} from "./qds-dialog-context.service"
 })
 export class DialogCloseButtonComponent extends CoreDialogCloseTriggerDirective {
   protected readonly qdsContext = useQdsDialogContext()
-  protected readonly buttonApi = useInlineIconButtonApi({
-    emphasis: "neutral",
+  protected readonly buttonApi = useIconButtonApi({
+    density: "compact",
     size: "md",
-    variant: "fixed",
+    variant: "ghost",
   })
 
   constructor() {