+++ packages/frameworks/angular-core/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @qualcomm-ui/angular-core Changelog
 
+## 2.4.8
+
+Aug 6th, 2026
+
+### Bug Fixes
+
+- [track-bindings]: handle spaces in class property ([46baae6](https://github.com/qualcomm/qualcomm-ui/commit/46baae6))
+
+## 2.4.7
+
+Jul 28th, 2026
+
+### Miscellaneous Chores
+
+- **deps:** update dependencies [@qualcomm-ui/core@1.11.7]
+
 ## 2.4.6
 
 Jul 1st, 2026